# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Control System API",
        "description": "API documentation for Risk Control System",
        "contact": {
            "email": "admin@riskcontrol.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/risk-rules": {
            "get": {
                "tags": [
                    "Risk Rules"
                ],
                "summary": "Get list of risk rules",
                "description": "Returns list of risk rules",
                "operationId": "getRiskRulesList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Risk Rules"
                ],
                "summary": "Store new risk rule",
                "description": "Returns risk rule data",
                "operationId": "storeRiskRule",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRiskRuleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/v1/risk-rules/{id}": {
            "get": {
                "tags": [
                    "Risk Rules"
                ],
                "summary": "Get risk rule information",
                "description": "Returns risk rule data",
                "operationId": "getRiskRuleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Risk Rule ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Risk Rules"
                ],
                "summary": "Update existing risk rule",
                "description": "Returns updated risk rule data",
                "operationId": "updateRiskRule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Risk Rule ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRiskRuleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Risk Rules"
                ],
                "summary": "Delete existing risk rule",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteRiskRule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Risk Rule ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/risk-rules/{id}/actions": {
            "post": {
                "tags": [
                    "Risk Rules"
                ],
                "summary": "Attach actions to risk rule",
                "description": "Returns risk rule data with attached actions",
                "operationId": "attachActionsToRiskRule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Risk Rule ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AttachActionsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/v1/risk-rules/actions": {
            "get": {
                "tags": [
                    "Risk Rules"
                ],
                "summary": "Get list of available actions",
                "description": "Returns list of configured actions",
                "operationId": "getAvailableActions",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AttachActionsRequest": {
                "title": "Attach Actions Request",
                "description": "Attach Actions request body data",
                "required": [
                    "action_ids"
                ],
                "properties": {
                    "action_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
...
            "StoreRiskRuleRequest": {
                "title": "Store Risk Rule Request",
                "description": "Store Risk Rule request body data",
                "required": [
                    "name",
                    "rule_type",
                    "severity",
                    "incident_limit"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Max Daily Loss"
                    },
                    "description": {
                        "type": "string",
                        "example": "Limits daily loss to $500"
                    },
                    "rule_type": {
                        "type": "string",
                        "enum": [
                            "min_duration",
                            "volume_consistency",
                            "trade_frequency"
                        ],
                        "example": "min_duration"
                    },
                    "severity": {
                        "type": "string",
                        "enum": [
                            "HARD",
                            "SOFT"
                        ],
                        "example": "HARD"
                    },
                    "incident_limit": {
                        "type": "integer",
                        "example": 3
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "parameters": {
                        "type": "object",
                        "example": {
                            "min_duration_seconds": 60
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdateRiskRuleRequest": {
                "title": "Update Risk Rule Request",
                "description": "Update Risk Rule request body data",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Max Daily Loss Updated"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated description"
                    },
                    "severity": {
                        "type": "string",
                        "enum": [
                            "HARD",
                            "SOFT"
                        ],
                        "example": "SOFT"
                    },
                    "incident_limit": {
                        "type": "integer",
                        "example": 5
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": false
                    },
                    "parameters": {
                        "type": "object",
                        "example": {
                            "min_duration_seconds": 120
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}